requirements.txt
py-cord
black
python-dotenv
motor[srv]